some security improvements, added README.md

+++ music_directory/main_app/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.forms import UserCreationForm
@@ -10,29 +10,32 @@ from .forms import MusicForm, LoginForm, MusicAdditionalForm
 # Create your views here.
 
 def index(request):
-    musics = Music.objects.all()
+    musics = Music.objects.filter(user = request.user)
     form = MusicForm
     return render(request, 'index.html', { 'musics': musics, 'form': form })
 
 def detail(request, music_id):
-    if request.method == 'POST':
-        instance = Music.objects.get(id = music_id)
-        form = MusicAdditionalForm(request.POST, instance = instance)
-        form_n = MusicForm(request.POST, instance = instance)
-        if form.is_valid():
-                music = form.save(commit = False)
-                music.user = request.user
-                music.save()
-        elif form_n.is_valid():
-                music = form_n.save(commit = False)
-                music.user = request.user
-                music.save()
-        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-    else:   
-        music = Music.objects.get(id = music_id)
-        form = MusicAdditionalForm
-        form_n = MusicForm
-        return render(request, 'details.html', { 'music': music, 'form': form, 'form_n': form_n })
+    if request.user.id == Music.objects.get(id = music_id).user_id:
+        if request.method == 'POST':
+            instance = Music.objects.get(id = music_id)
+            form = MusicAdditionalForm(request.POST, instance = instance)
+            form_n = MusicForm(request.POST, instance = instance)
+            if form.is_valid():
+                    music = form.save(commit = False)
+                    music.user = request.user
+                    music.save()
+            elif form_n.is_valid():
+                    music = form_n.save(commit = False)
+                    music.user = request.user
+                    music.save()
+            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        else:   
+            music = Music.objects.get(id = music_id)
+            form = MusicAdditionalForm
+            form_n = MusicForm
+            return render(request, 'details.html', { 'music': music, 'form': form, 'form_n': form_n })
+    else:
+        raise Http404
 
 def post_music(request):
     form = MusicForm(request.POST)
@@ -43,9 +46,12 @@ def post_music(request):
     return HttpResponseRedirect('/')
 
 def del_music(request, music_id):
-    instance = Music.objects.get(id = music_id)
-    instance.delete()
-    return HttpResponseRedirect('/')
+    if request.user.id == Music.objects.get(id = music_id).user_id:
+        instance = Music.objects.get(id = music_id)
+        instance.delete()
+        return HttpResponseRedirect('/')
+    else:
+        raise Http404
 
 def profile(request, username):
     user = User.objects.get(username = username)
@@ -125,66 +131,67 @@ def add_to_fav(request):
     return HttpResponse(favorites)
 
 def favorites(request):
-    musics = Music.objects.filter(favorites = True)
+    user = request.user
+    musics = Music.objects.filter(favorites = True, user = user)
     return render(request, 'favorites.html', {'musics': musics})
 
 def order_by_year(request):
     order = request.GET.get('order', 'desc')
-    all_music = Music.objects.all()
+    my_music = Music.objects.filter(user = request.user)
     if(order == 'desc'):
-        musics = all_music.order_by('-year')
+        musics = my_music.order_by('-year')
     elif(order == 'asc'):
-        musics = all_music.order_by('year')
+        musics = my_music.order_by('year')
     return render(request, 'favorites.html', {'musics': musics, 'order': order})
 
 def order_by_artist(request):
     order = request.GET.get('order', 'desc')
-    all_music = Music.objects.all()
+    my_music = Music.objects.filter(user = request.user)
     if(order == 'desc'):
-        musics = all_music.order_by('-artist')
+        musics = my_music.order_by('-artist')
     elif(order == 'asc'):
-        musics = all_music.order_by('artist')
+        musics = my_music.order_by('artist')
     return render(request, 'favorites.html', {'musics': musics, 'order': order})
 
 def order_by_genre(request):
     order = request.GET.get('order', 'desc')
-    all_music = Music.objects.all()
+    my_music = Music.objects.filter(user = request.user)
     if(order == 'desc'):
-        musics = all_music.order_by('-genre')
+        musics = my_music.order_by('-genre')
     elif(order == 'asc'):
-        musics = all_music.order_by('genre')
+        musics = my_music.order_by('genre')
     return render(request, 'favorites.html', {'musics': musics, 'order': order})
 
 def order_by_album(request):
     order = request.GET.get('order', 'desc')
-    all_music = Music.objects.all()
+    my_music = Music.objects.filter(user = request.user)
     if(order == 'desc'):
-        musics = all_music.order_by('-album')
+        musics = my_music.order_by('-album')
     elif(order == 'asc'):
-        musics = all_music.order_by('album')
+        musics = my_music.order_by('album')
     return render(request, 'favorites.html', {'musics': musics, 'order': order})
 
 def order_by_title(request):
     order = request.GET.get('order', 'desc')
-    all_music = Music.objects.all()
+    my_music = Music.objects.filter(user = request.user)
     if(order == 'desc'):
-        musics = all_music.order_by('-title')
+        musics = my_music.order_by('-title')
     elif(order == 'asc'):
-        musics = all_music.order_by('title')
+        musics = my_music.order_by('title')
     return render(request, 'favorites.html', {'musics': musics, 'order': order})
 
 def filter_year(request, year):
-    musics = Music.objects.filter(year = year)
+    musics = Music.objects.filter(user = request.user, year = year)
     return render(request, 'filter.html', {'musics': musics})
     
 def filter_artist(request, artist):
-    musics = Music.objects.filter(artist = artist)
+    musics = Music.objects.filter(user = request.user, artist = artist)
     return render(request, 'filter.html', {'musics': musics})
 
 def filter_genre(request, genre):
-    musics = Music.objects.filter(genre = genre)
+    musics = Music.objects.filter(user = request.user, genre = genre)
     return render(request, 'filter.html', {'musics': musics})
 
 def filter_album(request, album):
-    musics = Music.objects.filter(album = album)
+    musics = Music.objects.filter(user = request.user, album = album)
     return render(request, 'filter.html', {'musics': musics})
